pyproject.toml
[project]
name = "nonebot-plugin-analysenumber"
version = "0.1.0"
description = "analyse numbers plugin for NoneBot2"
authors = [
    {name = "Akirami", email = "Akiramiaya@outlook.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.1", "nonebot-adapter-onebot>=2.2.3"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/A-kirami/nonebot-plugin-analysenumber"
Repository = "https://github.com/A-kirami/nonebot-plugin-analysenumber"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
